fix waiting for acapy
